+++ test/OpenTelemetry.Instrumentation.Grpc.Tests/GrpcTests.client.cs
@@ -14,10 +14,10 @@
 // limitations under the License.
 // </copyright>
 
-#if !NETFRAMEWORK
 using System;
 using System.Diagnostics;
 using System.Linq;
+using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Greet;
@@ -26,6 +26,7 @@ using Grpc.Net.Client;
 using Microsoft.AspNetCore.Http;
 using Moq;
 using OpenTelemetry.Context.Propagation;
+using OpenTelemetry.Instrumentation.Grpc.Tests.GrpcTestHelpers;
 using OpenTelemetry.Instrumentation.GrpcNetClient;
 using OpenTelemetry.Instrumentation.GrpcNetClient.Implementation;
 using OpenTelemetry.Trace;
@@ -45,8 +46,16 @@ namespace OpenTelemetry.Instrumentation.Grpc.Tests
         [InlineData("http://[::1]", false)]
         public void GrpcClientCallsAreCollectedSuccessfully(string baseAddress, bool shouldEnrich = true)
         {
-            var uri = new Uri($"{baseAddress}:{this.server.Port}");
+            var uri = new Uri($"{baseAddress}:1234");
             var uriHostNameType = Uri.CheckHostName(uri.Host);
+
+            var httpClient = ClientTestHelpers.CreateTestClient(async request =>
+            {
+                var streamContent = await ClientTestHelpers.CreateResponseContent(new HelloReply()).DefaultTimeout();
+                var response = ResponseUtils.CreateResponse(HttpStatusCode.OK, streamContent, grpcStatusCode: global::Grpc.Core.StatusCode.OK);
+                response.TrailingHeaders().Add("grpc-message", "value");
+                return response;
+            });
 
             var processor = new Mock<BaseProcessor<Activity>>();
 
@@ -66,7 +75,10 @@ namespace OpenTelemetry.Instrumentation.Grpc.Tests
                     .AddProcessor(processor.Object)
                     .Build())
             {
-                var channel = GrpcChannel.ForAddress(uri);
+                var channel = GrpcChannel.ForAddress(uri, new GrpcChannelOptions
+                {
+                    HttpClient = httpClient,
+                });
                 var client = new Greeter.GreeterClient(channel);
                 var rs = client.SayHello(new HelloRequest());
             }
@@ -105,6 +117,7 @@ namespace OpenTelemetry.Instrumentation.Grpc.Tests
             Assert.Equal(0, activity.GetTagValue(SemanticConventions.AttributeRpcGrpcStatusCode));
         }
 
+#if !NETFRAMEWORK
         [Theory]
         [InlineData(true)]
         [InlineData(false)]
@@ -428,6 +441,7 @@ namespace OpenTelemetry.Instrumentation.Grpc.Tests
                 }));
             }
         }
+#endif
 
         [Fact]
         public void Grpc_BadArgs()
@@ -469,4 +483,3 @@ namespace OpenTelemetry.Instrumentation.Grpc.Tests
         }
     }
 }
-#endif
